feat: remove redis from commons
