+++ packages/api/package.json
@@ -36,6 +36,7 @@
   },
   "devDependencies": {
     "@types/cors": "2.8.13",
-    "@types/express": "4.17.17"
+    "@types/express": "4.17.17",
+    "ioredis": "5.3.2"
   }
 }
