+++ packages/commons/src/index.ts
@@ -16,9 +16,9 @@ export * from './getBlockAtTimestamp';
 export * from './extendBalancesWithTrade';
 export * from './getTimeInYearsBetweenTimestamps';
 export * from './getTimestampInSeconds';
+export * from './getTrustedProxies';
 export * from './isNull';
 export * from './isUndefined';
 export * from './provider';
-export * from './redis';
 export * from './sleep';
 export * from './token';
